Fix(isyncrc): pinentry-tty freezes decryption pipe
When the initial invoke of pinentry-tty isn't in tty, autostarting
`~/.local/bin/mbs-cron` results in all GPG commands that requires a
symmetric passphrase hanging up for minutes whenever the agent doesn't
have cached that key, because the non-GUI pinentry is still waiting for
the passphrase input, since it's called from a autostart script instead
of terminal there's no way for user to input the passphrase. Setting
pinentry mode to loopback with --decryption-pipe is a workaround.